# openapi.yaml
openapi: 3.0.3
info:
  title: Taskryse API
  version: 1.0.0
  description: >
    Backend contract for Taskryse. Money in minor units (kobo/cents).
    Core rule: plans grant ELIGIBILITY only; earnings come from APPROVED work
    against FUNDED sponsor escrow; promo credits are non-withdrawable.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
paths:
...
  /withdrawals:
    post:
      summary: Withdraw. Requires PIN + OTP. Deduct withdrawable, apply fee. States processing->completed|failed_reversed.
      responses: { '201': { description: created }, '402': { description: insufficient withdrawable }, '403': { description: pin/otp failed } }
...
  /campaigns/{id}/fund: { post: { summary: Fund escrow, parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: { '200': { description: funded } } } }
  /campaigns/{id}/analytics: { get: { parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: { '200': { description: ok } } } }
...
  /admin/kyc/{id}/decision: { post: { parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: { '200': { description: audit-logged } } } }
...
  /admin/withdrawals/{id}/release: { post: { parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: { '200': { description: released } } } }
  /admin/withdrawals/{id}/hold: { post: { parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: { '200': { description: held } } } }
...
  /admin/fraud/{id}/freeze: { post: { parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: { '200': { description: frozen } } } }
...